requirements.txt
django
pytz
sqlparse
django_heroku
gunicorn
django-cors-headers
djangorestframework